Refactor: 상세 위젯 파일 분리 #135
